+++ src/main/java/fr/estecka/variantscit/format/properties/ItemComponentProperty.java
@@ -1,13 +1,23 @@
 package fr.estecka.variantscit.format.properties;
 
+import java.util.Optional;
+import com.mojang.datafixers.util.Pair;
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.DataResult;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import fr.estecka.variantscit.CodecUtil;
+import fr.estecka.variantscit.format.INbtInput;
+import fr.estecka.variantscit.format.IStringTransform;
 import fr.estecka.variantscit.format.NbtAdapter;
+import fr.estecka.variantscit.format.NbtPath;
 import net.minecraft.component.ComponentType;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtElement;
+import net.minecraft.nbt.NbtOps;
+import net.minecraft.nbt.NbtString;
 import net.minecraft.registry.Registries;
+import net.minecraft.util.Identifier;
 
 public record ItemComponentProperty(
 	ComponentType<?> componentType,
@@ -21,6 +31,11 @@ implements IStringProperty
 			NbtAdapter.MAPCODEC.forGetter(ItemComponentProperty::nbtAdapter)
 		)
 		.apply(builder, ItemComponentProperty::new)
+	);
+
+	static public final Codec<TransformableProperty<ItemComponentProperty>> MONOSTRING_DECODER = Codec.STRING.flatXmap(
+		ItemComponentProperty::MonostringParse,
+		__->DataResult.error(()->"Encoding not supported")
 	);
 
 	@Override
@@ -42,5 +57,35 @@ implements IStringProperty
 
 		String result = this.nbtAdapter.ResolveData(nbt);
 		return result;
+	}
+
+	static private DataResult<TransformableProperty<ItemComponentProperty>> MonostringParse(String input){
+		int pathLocation;
+		for (pathLocation = 0; pathLocation<input.length(); pathLocation++){
+			char c = input.charAt(pathLocation);
+			if (c == '.' || !Identifier.isCharValid(c))
+				break;
+		}
+
+		String s_component = input.substring(0, pathLocation);
+		String s_path      = input.substring(pathLocation);
+
+		var r_path = (!s_path.isEmpty()) ? NbtPath.Parse(s_path) : DataResult.success(NbtPath.IDENTITY);
+		var r_component = Registries.DATA_COMPONENT_TYPE.getCodec().decode(NbtOps.INSTANCE, NbtString.of(s_component)).map(Pair::getFirst);
+
+		if (r_path.isError())
+			return r_path.map(__->null);
+
+		if (r_component.isError())
+			return r_component.map(__->null);
+
+		NbtPath path = r_path.getOrThrow();
+		ComponentType<?> component = r_component.getOrThrow();
+
+		return DataResult.success(new TransformableProperty<>(
+			new ItemComponentProperty(component, new NbtAdapter(path, INbtInput.AUTO)),
+			IStringTransform.AUTO,
+			Optional.empty()
+		));
 	}
 }